refactor: login session extension (#2676)
### TL;DR

Enhanced the login session extension functionality and improved UI for remaining session time display.
This is a follow-up PR of #2551

### What changed?

- Updated `LoginSessionExtendButton` component to display remaining session time in a more user-friendly format (days, hours, minutes, seconds)
- Added a tooltip to show the purpose of the displayed time
- Implemented auto-reload functionality when the session expires
- Integrated the login session extension button into the WebUIHeader with dividers
- Added translations for "Time until auto logout" in multiple languages

### How to test?

1. Log in to the application
2. Locate the login session extension button in the header
3. Hover over the displayed time to see the tooltip
4. Observe the countdown of the remaining session time
5. Click the "Extend" button to prolong the session
6. Let the session expire to test the auto-reload functionality

### Why make this change?

This change improves user experience by providing clearer information about the remaining login session time and making it easier for users to extend their sessions. The added tooltip and more detailed time format help users understand the purpose of the displayed information, while the auto-reload feature ensures a seamless experience when sessions expire.
