+++ react/src/components/MainLayout/WebUIHeader.tsx
@@ -7,15 +7,15 @@ import {
 import { useScrollBreakPoint } from '../../hooks/useScrollBreackPoint';
 import BAINotificationButton from '../BAINotificationButton';
 import Flex, { FlexProps } from '../Flex';
+import LoginSessionExtendButton from '../LoginSessionExtendButton';
 import ProjectSelect from '../ProjectSelect';
-import TimeContainer from '../TimeContainer';
 import UserDropdownMenu from '../UserDropdownMenu';
 import WEBUIHelpButton from '../WEBUIHelpButton';
 import WebUIThemeToggleButton from '../WebUIThemeToggleButton';
 // @ts-ignore
 import rawCss from './WebUIHeader.css?raw';
 import { MenuOutlined } from '@ant-design/icons';
-import { theme, Button, Typography, Grid } from 'antd';
+import { theme, Button, Typography, Grid, Divider } from 'antd';
 import _ from 'lodash';
 import { Suspense, useState, useTransition } from 'react';
 import { useTranslation } from 'react-i18next';
@@ -117,7 +117,11 @@ const WebUIHeader: React.FC<WebUIHeaderProps> = ({
           <WEBUIHelpButton />
           {baiClient.supports('extend-login-session') &&
             baiClient._config.enableExtendLoginSession && (
-              <TimeContainer format="HH:mm:ss" />
+              <Suspense>
+                <Divider type="vertical" />
+                <LoginSessionExtendButton />
+                <Divider type="vertical" />
+              </Suspense>
             )}
           <UserDropdownMenu />
         </Flex>
