+++ resources/i18n/en.json
@@ -633,7 +633,8 @@
     "General": "General",
     "Disabled": "Disabled",
     "NSelected": "Selected {{count}} item(s)",
-    "Extend": "Extend"
+    "Extend": "Extend",
+    "RemainingLoginSessionTime": "Time until auto logout"
   },
   "credential": {
     "Permission": "Permission",
@@ -977,9 +978,9 @@
     "ModifyImageResourceLimitReinstallRequired": "<p>If you modify an image that has already been installed, you will need to reinstall the image.</p><p>Modifications will not be applied until the image is reinstalled. \n</p>",
     "NoImagesAreSelected": "No images are selected",
     "AlreadyInstalledImage": "Already installed image",
-    "InstalledImagesAreExcluded": "Installed images are excluded." ,
+    "InstalledImagesAreExcluded": "Installed images are excluded.",
     "CheckImageInstallation": "Installing an image?",
-    "SearchImages":"Search images"
+    "SearchImages": "Search images"
   },
   "resourcePreset": {
     "ResourcePresets": "Resource Presets",
