+++ resources/i18n/ko.json
@@ -620,7 +620,8 @@
     "General": "일반",
     "Disabled": "비활성화",
     "NSelected": "{{count}}개 선택됨",
-    "Extend": "연장"
+    "Extend": "연장",
+    "RemainingLoginSessionTime": "자동 로그아웃까지 남은 시간"
   },
   "credential": {
     "Permission": "권한",
@@ -962,11 +963,11 @@
     "DescImageResourceModified": "리소스 제한이 수정되었습니다.",
     "DescImagePortsModified": "이미지 포트가 수정되었습니다.",
     "ModifyImageResourceLimitReinstallRequired": "<p>이미 설치된 이미지를 수정하는 경우, 이미지 재설치가 필요합니다.</p><<p>수정내용은 이미지가 재설치되기 전까지 적용되지 않습니다. </p>",
-    "NoImagesAreSelected":"선택된 이미지가 없습니다.",
-    "AlreadyInstalledImage":"이미 설치된 이미지입니다.",
-    "InstalledImagesAreExcluded":"설치된 이미지는 제외되었습니다.",
-    "CheckImageInstallation":"다음 이미지를 설치하시겠습니까?",
-    "SearchImages":"이미지 검색"
+    "NoImagesAreSelected": "선택된 이미지가 없습니다.",
+    "AlreadyInstalledImage": "이미 설치된 이미지입니다.",
+    "InstalledImagesAreExcluded": "설치된 이미지는 제외되었습니다.",
+    "CheckImageInstallation": "다음 이미지를 설치하시겠습니까?",
+    "SearchImages": "이미지 검색"
   },
   "resourcePreset": {
     "ResourcePresets": "자원 프리셋",
