+++ resources/i18n/ru.json
@@ -508,7 +508,8 @@
     "General": "Общие сведения",
     "Disabled": "Инвалид",
     "NSelected": "Выбранный {{count}} элемент(ы)",
-    "Extend": "Продлевать"
+    "Extend": "Продлевать",
+    "RemainingLoginSessionTime": "Время до автоматического выхода из системы"
   },
   "credential": {
     "Permission": "Разрешение",
