+++ resources/i18n/zh-CN.json
@@ -508,7 +508,8 @@
     "General": "一般情况",
     "Disabled": "残疾",
     "NSelected": "已选择的 {{count}} 项目",
-    "Extend": "延长"
+    "Extend": "延长",
+    "RemainingLoginSessionTime": "自动注销前的时间"
   },
   "credential": {
     "Permission": "允许",
